+++ src/frontend/src/optimizer/rule/mod.rs
@@ -68,8 +68,8 @@ pub use push_calculation_of_join_rule::*;
 mod join_commute_rule;
 mod over_window_to_agg_and_join_rule;
 pub use over_window_to_agg_and_join_rule::*;
-mod over_window_split_by_window_rule;
-pub use over_window_split_by_window_rule::*;
+mod over_window_split_rule;
+pub use over_window_split_rule::*;
 mod over_window_to_topn_rule;
 pub use join_commute_rule::*;
 pub use over_window_to_topn_rule::*;
